Refactor client library into separate files, add JSON escaping
Updates the build to make thigns a bit more consistent now we're
adding this. Next step will be to make a proper published build.

Rather than depend on a full Scala JSON library, we use the (small)
org.json library to do correct JSON string escaping. This means users
won't get conflicting versions of Circe (for example).
